requirements.txt
pyyaml
playwright
openai
anthropic
python-dotenv
flask
google-genai
